+++ test/integration/getVersion.test.ts
@@ -21,7 +21,7 @@ describe('getVersion', () => {
     const {version, compatibility} = await ada.getVersion()
 
     expect(version.major).to.equal(7)
-    expect(version.minor).to.equal(0)
+    expect(version.minor).to.equal(1)
 
     expect(version.flags.isDebug).to.equal(true)
 
@@ -44,6 +44,7 @@ describe('getVersion', () => {
       supportsBabbage: true,
       supportsCIP36Vote: true,
       supportsConway: true,
+      supportsMessageSigning: true,
     })
   })
 })